+++ lib/api/_loaders/assertion-scheduler.js
@@ -13,7 +13,7 @@ class AssertionScheduler {
   }
 
   /**
-   * @param {AssertionLoader} instance
+   * @param {AssertionInstance} instance
    * @param {{rescheduleInterval, abortOnFailure, stackTrace, timeout, reporter}} opts
    */
   constructor(instance, opts = {}) {
