+++ lib/api/client-commands/end.js
@@ -1,5 +1,4 @@
 const EventEmitter = require('events');
-const {Logger, Screenshots} = require('../../utils');
 
 /**
  * Ends the session. Uses session protocol command.
@@ -20,30 +19,6 @@ class End extends EventEmitter {
     const client = this.client;
 
     if (this.api.sessionId) {
-      if (this.testFailuresExist() && this.shouldTakeScreenshot()) {
-        Logger.info(`Failures in "${this.api.currentTest.name}". Taking screenshot...`);
-
-        const fileNamePath = Screenshots.getFileName({
-          testSuite: this.api.currentTest.module,
-          testCase: this.api.currentTest.name
-        }, client.options.screenshots);
-
-        this.api.saveScreenshot(fileNamePath, (result, err) => {
-          if (!err && this.client.transport.isResultSuccess(result))  {
-            const assertions = this.api.currentTest.results.assertions || [];
-            if (assertions.length > 0) {
-              const currentAssertion = assertions[assertions.length-1];
-              if (currentAssertion) {
-                currentAssertion.screenshots = currentAssertion.screenshots || [];
-                currentAssertion.screenshots.push(fileNamePath);
-              }
-            }
-          } else {
-            Logger.warn('Error saving screenshot...', err || result);
-          }
-        });
-      }
-
       this.api.session('delete', result => {
         client.session.clearSession();
         client.setApiProperty('sessionId', null);
