+++ lib/testsuite/index.js
@@ -8,7 +8,7 @@ const NightwatchAssertError = require('../assertion').AssertionError;
 const SuiteRetries = require('./retries.js');
 const Nightwatch = require('../index.js');
 const Concurrency = require('../runner/concurrency/concurrency.js');
-const {Logger} = require('../utils');
+const {Logger, Screenshots} = require('../utils');
 
 class TestSuite {
   constructor({modulePath, modules, settings, argv, addtOpts = {}}) {
@@ -420,6 +420,37 @@ class TestSuite {
     return this.runCurrentTest(currentTestName);
   }
 
+  shouldTakeScreenshot() {
+    return this.api.options.screenshots.enabled && this.api.options.screenshots.on_failure;
+  }
+
+  takeScreenshot() {
+    return new Promise((resolve, ) => {
+      Logger.info(`Failures in "${this.api.currentTest.name}". Taking screenshot...`);
+      const fileNamePath = Screenshots.getFileName({
+        testSuite: this.api.currentTest.module,
+        testCase: this.api.currentTest.name
+      }, this.client.options.screenshots);
+
+      this.api.saveScreenshot(fileNamePath, (result, err) => {
+        if (!err && this.client.transport.isResultSuccess(result))  {
+          const assertions = this.api.currentTest.results.assertions || [];
+          if (assertions.length > 0) {
+            const currentAssertion = assertions[assertions.length-1];
+            if (currentAssertion) {
+              currentAssertion.screenshots = currentAssertion.screenshots || [];
+              currentAssertion.screenshots.push(fileNamePath);
+            }
+          }
+        } else {
+          Logger.warn('Error saving screenshot...', err || result);
+        }
+        resolve();
+      });
+    }); 
+  }
+
+
   shouldRetrySuite() {
     return !this.reporter.allTestsPassed && this.suiteRetries.shouldRetrySuite();
   }
@@ -452,7 +483,7 @@ class TestSuite {
       if (!isAssertionError(err)) {
         // registering non-assert errors
         this.reporter.registerTestError(err);
-
+        this.emptyQueue();
         throw err;
       }
 
@@ -473,9 +504,13 @@ class TestSuite {
         });
       }
 
+      if(this.shouldTakeScreenshot()){
+        await this.takeScreenshot();
+      }
+
       // clearing the queue here to avoid continuing with the rest of the testcase,
       // unless abortOnFailure is set to false
-      if (!isAssertionError(err) || err.abortOnFailure) {
+      if (err.abortOnFailure) {
         this.emptyQueue();
       }
